I changed the game logic to game.tsx / Implemented visualization when you strike or save / I created the function to identify when the game ends / Created a modal to return to the player creation page and start over
